+++ tests/mocks/api/slash.py
@@ -1,4 +1,5 @@
 from typing import Dict
+
 from tests.common import random_string
 
 
@@ -15,6 +16,6 @@ def build_mock_slash_resp(command: str = '/hello', text: str = '', channel_name:
         'command': command,
         'text': text,
         'api_app_id': random_string(n_chars=8).upper(),
-        'response_url': f'https://hooks.slack.com/commands/ALKDSJFE/239847938417/thisisatest',
+        'response_url': 'https://hooks.slack.com/commands/ALKDSJFE/239847938417/thisisatest',
         'trigger_id': 'f12348713.32498734.12o138473294729734979174984159'
     }